Split vbmctl main.go into separate files
Move command implementations from main.go into dedicated files per
concern: create.go, delete.go, status.go, config.go, version.go,
and common.go.
